MWPW-154480 Split global-navigation.test.js into multiple files (#2576)
Brings total time to run tests from 40s -> 19s on my machine.
