+++ test/blocks/global-navigation/gnav-promo.test.js
@@ -1,0 +1,59 @@
+/* eslint-disable no-restricted-syntax */
+import { expect } from '@esm-bundle/chai';
+import { createFullGlobalNavigation } from './test-utilities.js';
+import { toFragment } from '../../../libs/blocks/global-navigation/utilities/utilities.js';
+
+describe('Promo', () => {
+  before(() => {
+    document.head.innerHTML = `<link rel="icon" href="/libs/img/favicons/favicon.ico" size="any">
+    <script src="https://auth.services.adobe.com/imslib/imslib.min.js" type="javascript/blocked" data-loaded="true"></script>
+    <script src="https://stage.adobeccstatic.com/unav/1.1/UniversalNav.js" type="javascript/blocked" data-loaded="true"></script>
+    `;
+  });
+
+  it('doesn\'t exist if metadata is not defined', async () => {
+    const nav = await createFullGlobalNavigation();
+    expect(nav.block.querySelector('.aside.promobar')).to.equal(null);
+  });
+
+  it('doesn\'t exist if metadata is not referencing a fragment', async () => {
+    const wrongPromoMeta = toFragment`<meta name="gnav-promo-source" content="http://localhost:2000/path/to/promo">`;
+    document.head.append(wrongPromoMeta);
+    const nav = await createFullGlobalNavigation({ hasPromo: true });
+    expect(nav.block.classList.contains('has-promo')).to.be.false;
+    expect(nav.block.querySelector('.aside.promobar')).to.equal(null);
+    wrongPromoMeta.remove();
+  });
+
+  it('doesn\'t exist if fragment doesn\'t contain an aside block', async () => {
+    const promoMeta = toFragment`<meta name="gnav-promo-source" content="http://localhost:2000/fragments/wrong-promo-fragment">`;
+    document.head.append(promoMeta);
+    const nav = await createFullGlobalNavigation({ hasPromo: true });
+    expect(nav.block.classList.contains('has-promo')).to.be.false;
+    expect(nav.block.querySelector('.aside.promobar')).to.equal(null);
+    promoMeta.remove();
+  });
+
+  it('is available if set up correctly', async () => {
+    const promoMeta = toFragment`<meta name="gnav-promo-source" content="http://localhost:2000/fragments/correct-promo-fragment">`;
+    document.head.append(promoMeta);
+    const nav = await createFullGlobalNavigation({ hasPromo: true });
+    expect(nav.block.classList.contains('has-promo')).to.be.true;
+    const asideElem = nav.block.querySelector('.aside.promobar');
+    expect(asideElem).to.exist;
+    expect(asideElem.getAttribute('daa-lh')).to.equal('Promo');
+    asideElem.querySelectorAll('a').forEach((linkElem) => {
+      expect(linkElem.hasAttribute('daa-ll')).to.be.true;
+    });
+    promoMeta.remove();
+  });
+
+  it('doesn\'t exist on mobile', async () => {
+    const promoMeta = toFragment`<meta name="gnav-promo-source" content="http://localhost:2000/fragments/correct-promo-fragment">`;
+    document.head.append(promoMeta);
+    const nav = await createFullGlobalNavigation({ viewport: 'mobile', hasPromo: true });
+    expect(nav.block.classList.contains('has-promo')).to.be.false;
+    const asideElem = nav.block.querySelector('.aside.promobar');
+    expect(asideElem).to.not.exist;
+  });
+});
